Vote power ups, fixed zoomo spectator, QOL
The sneak bomb event for voting has been limited to a powerup drops system where the player can pickup an item randomly spawned and be given the ability to perform it. Additionally the zoomo spectator fly issue has been resolved using a run task later. Quality of life changes and improvements to code.