Test Resulting Changes
The first official test of the project has been conducted, with Zoomo and Gubgame being tested along with voting and ready check, feedback has been collected and changes have been made based on my personal notes. Changes are yet to be made for the collected feedback but the team has been informed of the details and changes will be made.